# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NODAQ API — AI-powered business management for French SMEs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /affaires:
    get:
      operationId: listAffaires
      tags: [affaires]
      summary: List affaires
      parameters:
        - in: query
          name: statut
          schema:
            type: string
        - in: query
          name: inclureArchivees
          schema:
            type: boolean
      responses:
        "200":
          description: List of affaires
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffaireList"
    post:
      operationId: createAffaire
      tags: [affaires]
      summary: Create an affaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffaireInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Affaire"
  /affaires/{id}:
    get:
      operationId: getAffaire
      tags: [affaires]
      summary: Get a single affaire
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Affaire detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Affaire"
        "404":
          description: Not found
    patch:
      operationId: updateAffaire
      tags: [affaires]
      summary: Update an affaire
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffaireUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Affaire"
        "404":
          description: Not found
    delete:
      operationId: deleteAffaire
      tags: [affaires]
      summary: Delete an affaire
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /contrats:
    get:
      operationId: listContrats
      tags: [contrats]
      summary: List recurring contracts
      responses:
        "200":
          description: List of contrats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contrat"
    post:
      operationId: createContrat
      tags: [contrats]
      summary: Create a contrat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContratInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contrat"
  /contrats/{id}:
    patch:
      operationId: updateContrat
      tags: [contrats]
      summary: Update a contrat
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContratUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contrat"
        "404":
          description: Not found
    delete:
      operationId: deleteContrat
      tags: [contrats]
      summary: Delete a contrat
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Factures / Invoices ───────────────────────────────────────────────────
...
  /classeur:
    get:
      operationId: listClasseur
      tags: [classeur]
      summary: List documents in classeur
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClasseurList"
    post:
      operationId: createClasseurDocument
      tags: [classeur]
      summary: Add a document to classeur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClasseurDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClasseurDocument"
  /classeur/{id}:
    delete:
      operationId: deleteClasseurDocument
      tags: [classeur]
      summary: Delete a document from classeur
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Échéances fiscales ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Affaire:
      type: object
      required: [id, label, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        reference:
          type: ["string", "null"]
        label:
          type: string
        clientName:
          type: ["string", "null"]
        status:
          type: string
          enum: [PROSPECT, DEVIS_ENVOYE, ACCEPTEE, EN_COURS, TERMINEE, PERDUE, ARCHIVEE]
        quotedAmountCents:
          type: ["number", "null"]
        invoicedAmountCents:
          type: ["number", "null"]
        marginCents:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AffaireList:
      type: object
      required: [affaires, total]
      properties:
        affaires:
          type: array
          items:
            $ref: "#/components/schemas/Affaire"
        total:
          type: number
        totalQuotedCents:
          type: number
    AffaireInput:
      type: object
      required: [label]
      properties:
        label:
          type: string
          minLength: 1
        clientName:
          type: string
        quotedAmountCents:
          type: number
        status:
          type: string
        notes:
          type: string
        startDate:
          type: string
          format: date
    AffaireUpdate:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        clientName:
          type: string
        status:
          type: string
        quotedAmountCents:
          type: number
        invoicedAmountCents:
          type: number
        marginCents:
          type: number
        notes:
          type: string
        startDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date
...
    Contrat:
      type: object
      required: [id, label, cadence, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        label:
          type: string
        clientName:
          type: ["string", "null"]
        cadence:
          type: string
          enum: [mensuel, trimestriel, semestriel, annuel]
        amountCents:
          type: ["number", "null"]
        status:
          type: string
          enum: [ACTIF, SUSPENDU, TERMINE]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        nextOccurrenceDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContratInput:
      type: object
      required: [label, cadence]
      properties:
        label:
          type: string
          minLength: 1
        clientName:
          type: string
        cadence:
          type: string
          enum: [mensuel, trimestriel, semestriel, annuel]
        amountCents:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
    ContratUpdate:
      type: object
      properties:
        label:
          type: string
        clientName:
          type: string
        cadence:
          type: string
        amountCents:
          type: number
        status:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string

    # ── Factures ──────────────────────────────────────────────────────────────
...
    ClasseurDocument:
      type: object
      required: [id, name, category, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [FACTURES, CONTRATS, DEVIS, DIVERS]
        size:
          type: ["number", "null"]
        mimeType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        affaireId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ClasseurList:
      type: object
      required: [documents, total]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/ClasseurDocument"
        total:
          type: number
    ClasseurDocumentInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum: [FACTURES, CONTRATS, DEVIS, DIVERS]
        size:
          type: number
        mimeType:
          type: string
        notes:
          type: string
        affaireId:
          type: string

    # ── Échéances fiscales ─────────────────────────────────────────────────────
...