# lib/api-zod/src/generated/types/classeurDocument.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NODAQ API — AI-powered business management for French SMEs
 * OpenAPI spec version: 0.1.0
 */
import type { ClasseurDocumentCategory } from './classeurDocumentCategory';

export interface ClasseurDocument {
  id: string;
  name: string;
  category: ClasseurDocumentCategory;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  affaireId?: string | null;
  createdAt: Date;
}