# lib/api-zod/src/generated/types/classeurList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NODAQ API — AI-powered business management for French SMEs
 * OpenAPI spec version: 0.1.0
 */
import type { ClasseurDocument } from './classeurDocument';

export interface ClasseurList {
  documents: ClasseurDocument[];
  total: number;
}